+++ src/headers/rjmc_inf.hpp
@@ -64,8 +64,8 @@ namespace rjmc_sero{
         RObject dataList;
         List dataListCPP;
 
-        double counterFuncEval, counterAccepted, counterPosterior ,counterAdaptive;
-        double counterNonAdaptive;
+        int counterFuncEval, counterAccepted, counterPosterior ,counterAdaptive;
+        int counterNonAdaptive;
         double proposedLogPosterior, alpha, covarMaxVal, covarInitVal, covarInitValAdapt;
 
         std::function<VectorXd(RObject)> samplePriorDistributions;
